Separate temporal logic operator definition and rules
