+++ src/temporal-logic-lib/examples/compound_state_machine/safety.rs
@@ -8,6 +8,7 @@ use crate::examples::compound_state_machine::{
 };
 use crate::pervasive::seq::*;
 use crate::temporal_logic::*;
+use crate::temporal_logic_rules::*;
 use builtin::*;
 use builtin_macros::*;
