+++ src/temporal-logic-lib/examples/concurrent_controller/safety.rs
@@ -5,6 +5,7 @@ use crate::common::*;
 use crate::examples::concurrent_controller::state_machine::*;
 use crate::pervasive::seq::*;
 use crate::temporal_logic::*;
+use crate::temporal_logic_rules::*;
 use builtin::*;
 use builtin_macros::*;
