+++ src/temporal-logic-lib/examples/simple_controller/safety.rs
@@ -3,6 +3,7 @@
 #![allow(unused_imports)]
 use crate::examples::simple_controller::state_machine::*;
 use crate::temporal_logic::*;
+use crate::temporal_logic_rules::*;
 use builtin::*;
 use builtin_macros::*;
